# docs/swagger.yaml
openapi: "3.0.0"
info:
  title: LAAWS Crawler Service API
  description: |
    An API for managing  crawler tasks
  contact:
    name: LOCKSS Support
    url: 'https://www.lockss.org/'
    email: lockss-support@lockss.org
  license:
    name: BSD-3-Clause
    url: 'https://www.lockss.org/support/open-source-license/'
  version: 1.0.0
servers:
  - url: 'https://laaws.lockss.org:443/'
paths:
  /crawlers/:
    summary: Access point to all crawlers.
    description: Allows access to supported and installed crawlers.
    get:
      tags:
        - crawlers
      summary: Get the list of supported crawlers.
      description: Return the list of supported crawlers.
      operationId: getCrawlers
      responses:
        '200':
          description: The crawler list.
          content:
            application/json:
              schema:
                description: A list of crawlers.
                type: object
                properties:
                  crawls:
                    description: An array of crawler names
                    type: array
                    items:
                      type: string
        '404':
          description: No Such Crawler
          content:
            application/json: {}
...
  '/crawlers/{crawler}':
    summary: Return information about installed crawler.
    description: Access to information related to a crawler with crawler.
    get:
      tags:
        - crawlers
      summary: Return information about a crawler.
      description: Get information related to a installed crawler.
      operationId: getCrawlerByName
      responses:
        '200':
          description: Crawler Info Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/crawlerInfo'
        '401':
          description: Access Denied.
        '404':
          description: Not Found
    parameters:
      - name: crawler
        in: path
        description: Identifier for the crawler
        required: true
        schema:
          type: string
...
components:
  schemas:
...
    crawlerInfo:
      description: Information about a specific crawler.
      required:
        - isEnabled
        - isRunning
      type: object
      properties:
        isEnabled:
          description: Is the crawler enabled
          type: boolean
        isRunning:
          description: Is the crawl starter running
          type: boolean
        infoMap:
          description: key value pairs specific providing configuration information.
          type: array
          items:
            type: string
...
  securitySchemes:
    basicAuth:
      type: http
      description: HTTP Basic Authentication. Works over `HTTP` and `HTTPS`
      scheme: basic
security:
  - basicAuth: []
...