# src/generated/java/org/lockss/laaws/crawler/api/CrawlersApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.2).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package org.lockss.laaws.crawler.api;

import org.lockss.laaws.crawler.model.CrawlerInfo;
import org.lockss.laaws.crawler.model.InlineResponse200;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@Api(value = "crawlers", description = "the crawlers API")
public interface CrawlersApi {

    CrawlersApiDelegate getDelegate();

    @ApiOperation(value = "Return information about a crawler.", nickname = "getCrawlerByName", notes = "Get information related to a installed crawler.", response = CrawlerInfo.class, authorizations = {
        @Authorization(value = "basicAuth")
    }, tags={ "crawlers", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Crawler Info Found", response = CrawlerInfo.class),
        @ApiResponse(code = 401, message = "Access Denied."),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/crawlers/{crawler}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<CrawlerInfo> getCrawlerByName(@ApiParam(value = "Identifier for the crawler",required=true) @PathVariable("crawler") String crawler) {
        return getDelegate().getCrawlerByName(crawler);
    }


    @ApiOperation(value = "Get the list of supported crawlers.", nickname = "getCrawlers", notes = "Return the list of supported crawlers.", response = InlineResponse200.class, authorizations = {
        @Authorization(value = "basicAuth")
    }, tags={ "crawlers", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "The crawler list.", response = InlineResponse200.class),
        @ApiResponse(code = 404, message = "No Such Crawler") })
    @RequestMapping(value = "/crawlers/",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<InlineResponse200> getCrawlers() {
        return getDelegate().getCrawlers();
    }

}